+++ app/consumer/ante_handler.go
@@ -2,7 +2,7 @@ package app
 
 import (
 	errorsmod "cosmossdk.io/errors"
-	consumerante "github.com/allinbits/vaas/app/consumer/ante"
+	consumerante "github.com/allinbits/vaas/x/vaas/consumer/ante"
 	ibcconsumerkeeper "github.com/allinbits/vaas/x/vaas/consumer/keeper"
 	ibcante "github.com/cosmos/ibc-go/v10/modules/core/ante"
 	ibckeeper "github.com/cosmos/ibc-go/v10/modules/core/keeper"
